#257 [refactor] 이미지 저장 로직 모듈 분리
