+++ app/src/main/java/hous/release/android/presentation/personality/result/PersonalityResultViewModel.kt
@@ -1,13 +1,12 @@
 package hous.release.android.presentation.personality.result
 
-import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import hous.release.android.util.ImageDownloader
 import hous.release.android.util.UiEvent
 import hous.release.domain.entity.HomyType
 import hous.release.domain.entity.response.PersonalityResult
+import hous.release.domain.repository.ImageRepository
 import hous.release.domain.usecase.GetPersonalityResultUseCase
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -19,7 +18,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class PersonalityResultViewModel @Inject constructor(
-    private val application: Application,
+    private val imageRepository: ImageRepository,
     private val getPersonalityResultUseCase: GetPersonalityResultUseCase
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(PersonalityResult())
@@ -27,10 +26,6 @@ class PersonalityResultViewModel @Inject constructor(
 
     private val _uiEvent = MutableSharedFlow<UiEvent>()
     val uiEvent = _uiEvent.asSharedFlow()
-
-    private val imageDownloader: ImageDownloader by lazy {
-        ImageDownloader(application)
-    }
 
     fun getPersonalityResult(personalityColor: String) {
         viewModelScope.launch {
@@ -47,11 +42,11 @@ class PersonalityResultViewModel @Inject constructor(
         viewModelScope.launch {
             _uiEvent.emit(UiEvent.LOADING)
             try {
-                imageDownloader.downloadImage(
+                imageRepository.downloadImage(
                     _uiState.value.firstDownloadImageUrl,
                     getSaveImageFileName(_uiState.value.color, 1)
                 )
-                imageDownloader.downloadImage(
+                imageRepository.downloadImage(
                     _uiState.value.secondDownloadImageUrl,
                     getSaveImageFileName(_uiState.value.color, 2)
                 )
