+++ domain/src/main/java/hous/release/domain/repository/ImageRepository.kt
@@ -1,0 +1,5 @@
+package hous.release.domain.repository
+
+interface ImageRepository {
+    suspend fun downloadImage(imageUrl: String, fileName: String)
+}
